refactor(rocr): Move hsakmt_safe_env_to_int into libhsakmt.h
Drop hsakmt_env.h and define the shared env-var parser inline in
libhsakmt.h for Linux, virtio, and DXG openclose paths.
